Metropolitan Museum of Art support
Metropolitan Museum of Art support

+++ nga.py
@@ -10,21 +10,26 @@ import requests
 from bs4 import BeautifulSoup
 
 from a_constants import *
-from open_sqlite_handling import remove_old_and_create_new_sqlite_db
 from csv_info import get_file_creation_date
 from csv_info import get_month_name_from_int
 from a_constants import Constants as Const
 
 from open_source import Source
+from nga_open_object import Object
 
 
 class NGA(Source):
+    
+    def __init__(self):
+        self.objects = []
+        
+        # TODO: remove_old_and_create_new_sqlite_db()
 
     def download_open_data(self):
         self.download_zip_if_needed()
 
-    def unpack_and_recreate_db(self):
-        self.extract_zip_and_recreate_db()
+    def unpack_and_create_csv(self):
+        self.extract_zip_and_fix()
 
     @staticmethod
     def get_nga_latest_remote_git_update_date():
@@ -51,7 +56,8 @@ class NGA(Source):
         git_date = self.get_nga_latest_remote_git_update_date()
         file_date = get_file_creation_date(Constants.NGA_ZIP_FILE_PATH)
         if git_date > file_date:
-            print("A seemingly newer version is available for " + Const.NGA_REMOTE_DATA_ZIP)
+            print("A seemingly newer version is available for " +
+                  Const.NGA_REMOTE_DATA_ZIP)
         else:
             print(
                 "Your file is up to date. No need to download again from\n",
@@ -61,25 +67,18 @@ class NGA(Source):
         print(
             "File on server was uploaded:",
             git_date.year,
-            get_month_name_from_int(Constants.months_abbreviated, git_date.month),
+            get_month_name_from_int(
+                Constants.months_abbreviated, git_date.month),
             git_date.day,
             "\nYour downloaded local file ("
             + Constants.DOWNLOAD_FOLDER
             + Const.NGA_ZIP_FILE_NAME
             + ") is from:",
             file_date.year,
-            get_month_name_from_int(Constants.months_abbreviated, file_date.month),
+            get_month_name_from_int(
+                Constants.months_abbreviated, file_date.month),
             file_date.day,
         )
-
-    def extract_zip_and_recreate_db(self):
-        if os.path.exists(Constants.NGA_ZIP_FILE_PATH):
-            print(Constants.NGA_ZIP_FILE_PATH)
-            if os.path.exists(Const.NGA_FOLDER_RENAME_TO):
-                shutil.rmtree(Const.NGA_FOLDER_RENAME_TO)
-            self.extract_zip(Constants.NGA_ZIP_FILE_PATH, Const.NGA_CSV_DIR)
-            # os.rename("downloads/opendata-main.zip", "downloads/old/opendata-main.zip")
-        remove_old_and_create_new_sqlite_db()
 
     def download_zip_if_needed(self):
         if os.path.exists(Constants.NGA_ZIP_FILE_PATH):
@@ -95,21 +94,100 @@ class NGA(Source):
                 if input(
                         f"Extract the downloaded {Const.NGA_ZIP_FILE_NAME} \
                         and create/replace the database? \n yes/no ") == "yes":
-                    self.extract_zip_and_recreate_db()
+                    self.extract_zip_and_fix()
         else:
             print("No file found. Downloading...")
             self.download_zip(
                 Const.NGA_REMOTE_DATA_ZIP, Constants.DOWNLOAD_FOLDER
             )
-            self.extract_zip_and_recreate_db()
+            self.extract_zip_and_fix()
 
     @staticmethod
-    def drop_unwanted_columns(df):
+    def drop_unwanted_columns(df: pd.DataFrame):
         for col in df.columns:
             if col not in Const.COLUMNS_USED:
                 df.drop(col, axis=1, inplace=True)
         return df
 
+    
+
+    #def add_fixed_urls(self, df):
+        
+        #temp = [] 
+        
+        #for index, row in df.iterrows():
+            
+            
+    
+            #return df
+    
+        """ nga_object = Object(row["source"][index],
+        row["objectid"][index],
+        row["title"][index],
+        row["attribution"][index],
+        row["beginyear"][index],
+        row["endyear"][index],
+        row["displaydate"][index],
+        row["classification"][index],
+        row["medium"][index],
+        row["width"][index],
+        row["height"][index], 
+        row["iiifurl"][index]) """
+        
+        #temp.append(nga_object)
+
+        # Generated inside Object constructor
+        """ full_url
+        imgurl_downsized
+        imgurl_thumb """
+        
+        
+    def fix_image_properties(self, df: pd.DataFrame):
+    
+        for index, row in df.iterrows():
+            # print(row["width"][index])
+            width = int(df["width"][index])
+            height = int(df["height"][index])
+            iiifurl = str(df["iiifurl"][index])
+            
+            # print("width etc ", width, height, iiifurl)
+            
+            imgurl_full = ""
+            imgurl_downsized = ""
+            imgurl_thumb = ""
+
+            if width > 4096 or height > 4096:
+                if width > 4096:
+                    width = 4096
+
+                if height > 4096:
+                    height = 4096
+
+                imgurl_full = f"{iiifurl}/full/!{str(width)},{str(height)}/0/default.jpg"
+            else:
+                imgurl_full = f"{iiifurl}/full/{str(width)},{str(height)}/0/default.jpg"
+
+            if height > 1500 or width > 1500:
+                imgurl_downsized = iiifurl + "/full/!1500,1500/0/default.jpg"
+            else:
+                imgurl_downsized = iiifurl + f"/full/!{str(width)},{str(height)}/0/default.jpg"
+            
+            imgurl_thumb = iiifurl + "/full/!200,200/0/default.jpg"
+                
+            df.at[index, "source"] =  "NGA"
+            df.at[index, "imgurl_thumb"]=  imgurl_thumb
+            df.at[index, "imgurl_downsized"]= imgurl_downsized
+            df.at[index, "imgurl_full"] = imgurl_full
+            
+            #df.get_index()
+            #df.insert(0, 'source', 'NGA')
+            
+            #df.assign('full_urlthumb', imgurl_thumb)
+            #df.assign('full_urldownsized', imgurl_downsized)
+            #df.assign('full_url', full_url)
+            
+        return df
+    
     def merge(self, csv1_file_name, csv2_file_name):
         # reading two csv files
         df1 = pd.read_csv(csv1_file_name, on_bad_lines='skip', index_col=False, dtype='unicode')
@@ -121,11 +199,16 @@ class NGA(Source):
         res = pd.merge(df1, df2,
                        on=['objectid'],
                        how='inner')
-
-        res.insert(0, 'source', 'NGA')
+        
+        res = self.fix_image_properties(res)  # fix_image_properties()
+        
+        res.drop('iiifurl', inplace=True, axis=1)
+        
         res.to_csv(csv1_file_name, index=False)
         if os.path.exists(csv2_file_name):
             os.remove(csv2_file_name)
+            
+        return csv1_file_name
 
     def fix_nga_csv_in_folder(self, art_folder):
         csv_file_names = glob.glob(os.path.join(art_folder, "*.csv"))
@@ -152,27 +235,33 @@ class NGA(Source):
         self.merge(Const.NGA_CSV_CONTAINER + "\\objects.csv",
                    Const.NGA_CSV_CONTAINER + "\\published_images.csv")
 
-    def extract_zip(self, zip_name, extract_to):
-        with ZipFile(zip_name, "r") as zip_f:
-            zip_f.printdir()
-            print("Extract start...")
+    def extract_zip_and_fix(self):
+        
+        if os.path.exists(Constants.NGA_ZIP_FILE_PATH):
+            print(Constants.NGA_ZIP_FILE_PATH)
+            if os.path.exists(Const.NGA_FOLDER_RENAME_TO):
+                shutil.rmtree(Const.NGA_FOLDER_RENAME_TO)
+            
+            with ZipFile(Constants.NGA_ZIP_FILE_PATH, "r") as zip_f:
+                zip_f.printdir()
+                print("Extract start...")
 
-            for file in zip_f.namelist():
-                if file.startswith(Const.NGA_DOWNLOAD_STARTS_WITH):
+                for file in zip_f.namelist():
+                    if file.startswith(Const.NGA_DOWNLOAD_STARTS_WITH):
 
-                    end_part = file.split("data/")
-                    end_part = end_part[len(end_part) - 1]
-                    # print(end_part)
-                    if end_part in Const.FILES_USED:
-                        zip_f.extract(file, extract_to)
+                        end_part = file.split("data/")
+                        end_part = end_part[len(end_part) - 1]
+                        # print(end_part)
+                        if end_part in Const.FILES_USED:
+                            zip_f.extract(file, Const.NGA_FOLDER_RENAME_TO)
 
-            if zip_f.namelist():
-                os.rename(Const.NGA_CSV_DIR + Const.NGA_FOLDER_TO_RENAME, Const.NGA_FOLDER_RENAME_TO)
-                print("Now cleaning up the files in", Const.NGA_CSV_CONTAINER)
-                self.fix_nga_csv_in_folder(Const.NGA_CSV_CONTAINER)
-                print("Extract complete!")
-            else:
-                print("Extract failed.")
+                if zip_f.namelist():
+                    os.rename(Const.NGA_CSV_DIR + Const.NGA_FOLDER_TO_RENAME, Const.NGA_FOLDER_RENAME_TO)
+                    print("Now cleaning up the files in", Const.NGA_CSV_CONTAINER)
+                    self.fix_nga_csv_in_folder(Const.NGA_CSV_CONTAINER)
+                    print("Extract complete!")
+                else:
+                    print("Extract failed.")
 
     @staticmethod
     def download_zip(data_url, download_folder):
